Session

+++ Controllers/StockController.cs
@@ -1,6 +1,7 @@
 ﻿using aksjehandel.DAL;
 using aksjehandel.Models;
 using Castle.Core.Logging;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
@@ -20,6 +21,7 @@ namespace aksjehandel.Controllers
         private readonly IStockRepository _db;
 
         private ILogger<StockController> _log;
+        private const string _signedIn = "signedIn";
         public StockController(IStockRepository db, ILogger<StockController> log)
         {
             _db = db;
@@ -27,6 +29,10 @@ namespace aksjehandel.Controllers
         }
         public async Task<ActionResult> regOrder(Order newOrder)
         {
+            if (string.IsNullOrEmpty(HttpContext.Session.GetString(_signedIn)))
+            {
+                return Unauthorized();
+            }
             if (ModelState.IsValid)
             {
                 bool returnOK = await _db.RegOrder(newOrder);
@@ -42,6 +48,10 @@ namespace aksjehandel.Controllers
         }
         public async Task<ActionResult> DeleteOrder(int id)
         {
+            if (string.IsNullOrEmpty(HttpContext.Session.GetString(_signedIn)))
+            {
+                return Unauthorized();
+            }
             bool returnOK = await _db.DeleteOrder(id);
             if (!returnOK)
             {
@@ -52,6 +62,10 @@ namespace aksjehandel.Controllers
         }
         public async Task<ActionResult> ChangeOrder(Order changeOrder)
         {
+            if (string.IsNullOrEmpty(HttpContext.Session.GetString(_signedIn)))
+            {
+                return Unauthorized();
+            }
             if (ModelState.IsValid)
             {
                 bool returnOk = await _db.ChangeOrder(changeOrder);
@@ -67,6 +81,10 @@ namespace aksjehandel.Controllers
         }
         public async Task<ActionResult> GetOneOrder(int id)
         {
+            if (string.IsNullOrEmpty(HttpContext.Session.GetString(_signedIn)))
+            {
+                return Unauthorized();
+            }
             Order oneOrder = await _db.GetOneOrder(id);
             if (oneOrder == null)
             {
@@ -77,28 +95,48 @@ namespace aksjehandel.Controllers
         }
         public async Task<ActionResult> getAllOrders(int portfolioId)
         {
+            if (string.IsNullOrEmpty(HttpContext.Session.GetString(_signedIn)))
+            {
+                return Unauthorized();
+            }
             List<Order> allOrders = await _db.GetAllOrders(portfolioId);
             return Ok(allOrders);
 
         }
         public async Task<ActionResult> GetAllShareholdings(int portfolioId)
         {
+            if (string.IsNullOrEmpty(HttpContext.Session.GetString(_signedIn)))
+            {
+                return Unauthorized();
+            }
             List<Shareholding> allShareholdings = await _db.GetAllShareholdings(portfolioId);
             return Ok(allShareholdings);
         }
         public async Task<ActionResult> GetAllPortfolios()
         {
+            if (string.IsNullOrEmpty(HttpContext.Session.GetString(_signedIn)))
+            {
+                return Unauthorized();
+            }
             List<Portfolio> allPortfolios = await _db.GetAllPortfolios();
             return Ok(allPortfolios);
         }
         public async Task<ActionResult> GetAllCompanies()
         {
+            if (string.IsNullOrEmpty(HttpContext.Session.GetString(_signedIn)))
+            {
+                return Unauthorized();
+            }
             List<Company> allCompanies = await _db.GetAllCompanies();
             return Ok(allCompanies);
 
         }
         public async Task<ActionResult> GetAllTrades()
         {
+            if (string.IsNullOrEmpty(HttpContext.Session.GetString(_signedIn)))
+            {
+                return Unauthorized();
+            }
             List<Trade> allTrades = await _db.GetAllTrades();
             return Ok(allTrades);
         }
@@ -111,12 +149,20 @@ namespace aksjehandel.Controllers
                 if (!returnOk)
                 {
                     _log.LogInformation("Innloggingen feilet for bruker. Brukernavn: " + user.Username);
+                    HttpContext.Session.SetString(_signedIn, "");
                     return Ok(false);
                 }
+                HttpContext.Session.SetString(_signedIn, "signedIn");
                 return Ok(true);
             }
             _log.LogInformation("Feil i inputvalidering");
             return BadRequest("Feil i inputvalidering på server");
         }
+
+        public void SignOut()
+        {
+            HttpContext.Session.SetString(_signedIn, "");
+        }
+
     }
 }
